refactor: kill dead code. refactor: improve file hierarchy
