+++ browser/index.js
@@ -7,7 +7,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import Simulation from './simulation'
 import SimulationContainer from './container/Simulation'
-import Menu from './menu'
+import Menu from './menu/menu'
 import Navigator from './navigator'
 import Intro from './intro.jsx'
 import introText from './introText.js'
