+++ browser/main.js
@@ -3,8 +3,8 @@ import React from 'react'
 import {Router, Route, IndexRedirect, browserHistory} from 'react-router'
 import {render} from 'react-dom'
 import {connect, Provider} from 'react-redux'
-import App from './index.js'
-import store from './store.jsx'
+import App from './index'
+import store from './store'
 
 render(<Provider store={store}>
         <App />
