+++ browser/navigator/navContent.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import {Row, Navbar, NavItem} from 'react-materialize'
-import Rules from '../rules.jsx'
-import Panels from '../navinstrucpanels.jsx'
-import {panel1, panel2, panel3} from '../panelData.js'
+import Rules from './rules'
+import Panels from './navinstrucpanels'
+import {panel1, panel2, panel3} from '../panelData'
 
 export default class NavigationBar extends React.Component {
   constructor(props) {
