+++ browser/navigator/rules.js
