+++ browser/tutorial/intro.js
@@ -3,13 +3,8 @@ import 'aframe-animation-component'
 import 'babel-polyfill'
 import { Entity, Scene } from 'aframe-react'
 import React from 'react'
-import Tutorial from './tutorial/tutorial'
+import Tutorial from './tutorial'
 import UserCam from './userCam'
-
-const driverIntro = `Hello space pilot ...\n
-We recieved your transmission for help from ground control. We'll patch you to an operator in a moment.\n
-You'll need to follow instructions from our operator to get your ship functional again. Stay calm and don't crash into the sun!\n
-*Click Text Anywhere To Continue*`
 
 export default class Intro extends React.Component {
   constructor(props) {
